pyproject.toml
[project]
name = "jsonvalidator"
version = "0.1.0"
description = ""
authors = [{ name = "Alanis Gonzalez, Jose Adrian" }]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
  "jsonschema (>=4.25.1,<5.0.0)",
  "loguru (>=0.7.3,<0.8.0)",
  "pyyaml (>=6.0.2,<7.0.0)",
  "genson (>=1.3.0,<2.0.0)",
]

[tool.poetry]
package-mode = true
packages = [{ include = "jsonvalidator" }]

[project.scripts]
runner = 'jsonvalidator.validator:run'
generate_schema = 'jsonvalidator.validator:generate_schema'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
